move prepare into dev
